+++ .spin/cmds.py
@@ -1,8 +1,9 @@
+import os
 import site
 import subprocess
 import webbrowser
 from pathlib import Path
-import os
+
 import click
 from spin.cmds import meson
